Principalの条件をtodo update deleteの時に投げる

+++ src/main/java/prv/koplec/jaxrs_todo/resource/TodoResource.java
@@ -2,17 +2,27 @@ package prv.koplec.jaxrs_todo.resource;
 
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.SecurityContext;
 import prv.koplec.jaxrs_todo.entities.Todo;
+import prv.koplec.jaxrs_todo.entities.User;
 import prv.koplec.jaxrs_todo.services.TodoService;
 import prv.koplec.jaxrs_todo.services.TodoServiceImpl;
+import prv.koplec.jaxrs_todo.services.UserService;
+import prv.koplec.jaxrs_todo.services.UserServiceImpl;
 
+import java.security.Principal;
 import java.util.List;
 
 @Path("/todos")
 public class TodoResource {
     private final TodoService todoService;
+    private final UserService userService;
+
+    @Context
+    private SecurityContext securityContext;
 
     // @Inject
     // public TodoResource(TodoService todoService) {
@@ -21,6 +31,7 @@ public class TodoResource {
 
     public TodoResource() {
         this.todoService = new TodoServiceImpl();
+        this.userService = new UserServiceImpl();
     }
 
     @GET
@@ -45,6 +56,7 @@ public class TodoResource {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
+    // todoはほかのユーザIDのものを作ってもよい
     public Response createTodo(Todo todo) {
         Todo createdTodo = todoService.createTodo(todo);
         return Response.status(Response.Status.CREATED).entity(createdTodo).build();
@@ -55,24 +67,56 @@ public class TodoResource {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     public Response updateTodo(@PathParam("id") Long id, Todo updatedTodo) {
-        Todo existingTodo = todoService.updateTodo(id, updatedTodo);
-        if (existingTodo != null) {
-            return Response.ok(existingTodo).build();
+        //ログインユーザの情報取得
+        Principal principal = securityContext.getUserPrincipal();
+        
+        if (principal != null) {
+            String username = principal.getName();
+            User principalUser = userService.getUserByUsername(username);
+            // ユーザが自分のTodoを更新しようとしているか確認
+            Todo existingTodo = todoService.getTodoById(id);
+            if(existingTodo == null){
+                return Response.status(Response.Status.NOT_FOUND).entity("Todo not found").build(); 
+            }
+            Long userId = existingTodo.getUserId();
+            if(!principalUser.getId().equals(userId)){
+                return Response.status(Response.Status.FORBIDDEN).entity("You don't have permission to update this todo").build();
+            }
+
+            Todo updated = todoService.updateTodo(id, updatedTodo);
+            return Response.ok(updated).build();
         } else {
-            return Response.status(Response.Status.NOT_FOUND).entity("Todo not found").build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity("User not authenticated").build();
         }
     }
 
     @DELETE
     @Path("/{id}")
     public Response deleteTodo(@PathParam("id") Long id) {
-        boolean deleted  = todoService.deleteTodo(id);
-        if (deleted) {
-            return Response.noContent().build();
-        } else {
-            return Response.status(Response.Status.NOT_FOUND).entity("Todo not found").build();
+        //ログインユーザの情報取得
+        Principal principal = securityContext.getUserPrincipal();
+        
+        if (principal != null) {
+            String username = principal.getName();
+            User principalUser = userService.getUserByUsername(username);
+            // ユーザが自分のTodoを更新しようとしているか確認
+            Todo existingTodo = todoService.getTodoById(id);
+            if(existingTodo == null){
+                return Response.status(Response.Status.NOT_FOUND).entity("Todo not found").build(); 
+            }
+            Long userId = existingTodo.getUserId();
+            if(!principalUser.getId().equals(userId)){
+                return Response.status(Response.Status.FORBIDDEN).entity("You don't have permission to update this todo").build();
+            }
+
+            boolean deleted  = todoService.deleteTodo(id);
+            if (deleted) {
+                return Response.noContent().build();
+            } else {
+                return Response.status(Response.Status.NOT_FOUND).entity("Todo not found").build();
+            }
+        }else{
+            return Response.status(Response.Status.UNAUTHORIZED).entity("User not authenticated").build();
         }
     }
-
-
 }
